Issue #2632636: Moved forms and controllers out of Entity namespace
